+++ core/common/src/main/kotlin/cn/wj/android/cashbook/core/common/ext/Number.kt
@@ -18,26 +18,32 @@ package cn.wj.android.cashbook.core.common.ext
 
 import java.math.BigDecimal
 
+/** 将 [String] 类型数字转换为 [BigDecimal]，为空或转换失败，返回值为 `0` 的 [BigDecimal] */
 fun String?.toBigDecimalOrZero(): BigDecimal {
     return this?.toBigDecimalOrNull() ?: "0".toBigDecimal()
 }
 
+/** 将 [Number] 类型数字转换为 [BigDecimal]，为空或转换失败，返回值为 `0` 的 [BigDecimal] */
 fun Number?.toBigDecimalOrZero(): BigDecimal {
     return this?.toString().toBigDecimalOrZero()
 }
 
+/** 将 [String] 类型数字转换为 [Float]，为空或转换失败，返回值为 `0f` 的 [Float] */
 fun String?.toFloatOrZero(): Float {
     return this?.toFloatOrNull() ?: 0f
 }
 
+/** 将 [String] 类型数字转换为 [Double]，为空或转换失败，返回值为 `0.0` 的 [Double] */
 fun String?.toDoubleOrZero(): Double {
     return this?.toDoubleOrNull() ?: 0.0
 }
 
+/** 将 [String] 类型数字转换为 [Int]，为空或转换失败，返回值为 `0` 的 [Int] */
 fun String?.toIntOrZero(): Int {
     return this?.toIntOrNull() ?: 0
 }
 
+/** 将 [Int] 类型的数字转换为 [String]，并在数字小于 10 时，在前补 0 */
 fun Int.completeZero(): String {
     return if (this >= 10) {
         this.toString()