+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 # 记账本 Changelog
 
 ## [Unreleased]
+### Add
+- 账本添加自定义背景功能；
+### Fixed
+- 修复宽屏设备下，首页图片显示不完整问题；
+### Changed
+- 将DatePicker、TimePicker替换为Compose版本实现;
 
 ## [1.0.6_24082118]
 ### Add